Allow replacing plugins of same type/user; allow custom contexts
Replacing plugins is necessary for runtime updates.

Custom contexts is necessary for a plugin to use a wide range of Camel
features. You could argue actually that the only thing a plugin need to
do is hand us a context, and it's responsible for adding routes to it. I
think we will probably move this way.